Fix UnboundLocalError in transcribe_audio function
Added global declaration for is_speech_synthesis_active at start
of transcribe_audio function.

Error was caused by setting the variable in the finalization block
without declaring it as global, creating a local variable that
shadowed the global and caused UnboundLocalError when reading it.

This fixes the STT crash on startup.